WIP: Add ToC and other missing start/end bits for individual pages
The XSLT process has a bunch of special cases for generating table of
contents, but upon inspection I suspect they are unnecessary; I'm
initially trying a much simpler approach that can apply equally to
both Techniques and Understanding.

ToC is not 100% correct in the new process yet due to headings not yet
being reordered the way they are in the previous process.
