+++ _includes/head.html
@@ -1,0 +1,8 @@
+{% # common tags inserted into all non-index/about pages %}
+<meta name="viewport" content="width=device-width, initial-scale=1"/>
+<link rel="stylesheet" href="https://w3.org/WAI/assets/css/style.css" />
+{% if isTechniques %}
+    {% include "techniques/head.html" %}
+{% elsif isUnderstanding %}
+    {% include "understanding/head.html" %}
+{% endif %}
