+++ _includes/sidebar.html
@@ -1,0 +1,13 @@
+<aside class="box nav-hack sidebar standalone-resource__sidebar" aria-labelledby="page-contents">
+    <nav>
+        <header class="box-h" id="page-contents">Page Contents</header>
+        <div class="box-i">
+            <ul>
+            {% # tocLinks are populated in eleventy config %}
+            {% for link in tocLinks %}
+                <li><a href="{{ link.href }}">{{ link.label }}</a></li>
+            {% endfor %}
+            </ul>
+        </div>
+    </nav>
+</aside>
