+++ _includes/understanding/head.html
@@ -1,3 +1,1 @@
-{% # common tags inserted into all non-index/about pages %}
-<link rel="stylesheet" href="https://w3.org/WAI/assets/css/style.css" />
 <link rel="stylesheet" href="base.css" />
