+++ _includes/waiscript.html
@@ -8,7 +8,7 @@
 	var translationStrings = {}; /* fix WAI JS */
 </script>
 <script src="https://www.w3.org/WAI/assets/scripts/main.js"></script>
-<!-- Matomo -->
+{% # Matomo %}
 <script>
 	var _paq = _paq || [];
 	/* tracker methods like "setCustomDimension" should be called before "trackPageView" */
@@ -26,4 +26,4 @@
 <noscript>
 	<p><img src="//www.w3.org/analytics/piwik/piwik.php?idsite=328&amp;rec=1" style="border:0;" alt="" /></p>
 </noscript>
-<!-- End Matomo Code -->
+{% # End Matomo Code %}
